add finalizer and predicate for status

+++ internal/controllers/s3bucket/cleaner.go
@@ -6,14 +6,17 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/opdev/subreconciler"
+	"github.com/snapp-incubator/s3-operator/pkg/consts"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 // Cleanup cleans up the provisioned resources for the s3Bucket object
 func (r *Reconciler) Cleanup(ctx context.Context) (ctrl.Result, error) {
 	// Do the actual reconcile work
 	subrecs := []subreconciler.Fn{
-		r.removeBucket,
+		r.removeOrRetainBucket,
+		r.removeCleanupFinalizer,
 	}
 	for _, subrec := range subrecs {
 		result, err := subrec(ctx)
@@ -25,7 +28,11 @@ func (r *Reconciler) Cleanup(ctx context.Context) (ctrl.Result, error) {
 	return subreconciler.Evaluate(subreconciler.DoNotRequeue())
 }
 
-func (r *Reconciler) removeBucket(ctx context.Context) (*ctrl.Result, error) {
+func (r *Reconciler) removeOrRetainBucket(ctx context.Context) (*ctrl.Result, error) {
+	// Clean only if deletionPolicy is on Delete mode
+	if r.s3Bucket.Spec.S3DeletionPolicy == consts.DeletionPolicyRetain {
+		return subreconciler.ContinueReconciling()
+	}
 	err := r.s3Agent.deleteBucket(r.s3BucketName)
 	if err != nil {
 		if aerr, ok := err.(awserr.Error); ok {
@@ -37,6 +44,20 @@ func (r *Reconciler) removeBucket(ctx context.Context) (*ctrl.Result, error) {
 		}
 		r.logger.Error(err, "failed to remove the bucket")
 		return subreconciler.Requeue()
+	}
+	return subreconciler.ContinueReconciling()
+}
+
+func (r *Reconciler) removeCleanupFinalizer(ctx context.Context) (*ctrl.Result, error) {
+	if r.s3Bucket == nil {
+		return subreconciler.ContinueReconciling()
+	}
+
+	if objUpdated := controllerutil.RemoveFinalizer(r.s3Bucket, consts.S3BucketCleanupFinalizer); objUpdated {
+		if err := r.Update(ctx, r.s3Bucket); err != nil {
+			r.logger.Error(err, "failed to remove finalizer from s3Bucket")
+			return subreconciler.Requeue()
+		}
 	}
 	return subreconciler.ContinueReconciling()
 }
